Update with all the changes in core libraries since I last worked on this. Program compiles but there are bugs now.

+++ test/test_program_options.cpp
@@ -27,7 +27,7 @@ namespace BoostPOW {
     struct test_case {
         bool ExpectValid;
         stack<string> Input;
-        int ArgCount;
+        size_t ArgCount;
         char ** ArgValues;
 
         test_case (bool expect_valid, stack<string> in) :
